Adds the api_keys config option for multiple geocoders
For the use case in which we'd like to use different geocoders which
require api keys, I added a new option that takes a Hash of the form
`{service => api_key, ...}`.

I do feel, however, that this might be telling of my non-standard use of
the gem: I use the `Geocoder::search` method directly and with various
geocoders instead of using it for a rails model automagically with a
single geocoder. I just really like the code organization and
extensibility of this gem :) 